fix(Steps): story
